Implementar búsqueda por filtros

+++ app/src/main/java/edu/iesam/bikerly/data/local/mock/Mappers.kt
@@ -8,9 +8,9 @@ fun MotorbikeMockModel.toModel(): Motorbike {
         this.id,
         this.make,
         this.model,
-        this.year,
+        this.year.toInt(),
         this.type,
-        this.displacement,
+        this.displacement.toInt(),
         this.img.toUri()
     )
 }
